refactor(health): rename `NgIndicators` to `HealthIndicators`
Since the Indicators also include fixing indicators, so rename it to HealthIndicators.
