+++ packages/cube-frontend-web-app/src/pages/home/health/[module]/HealthTimeBar/HealthIndicators.tsx
